Diverge and commit
It is known that hashes may collide. When they collide the program would
just crash.
By accepting divergence nodes with identical IDs may actually represent
different structure.
Even without me explicitly programming anything about it, it may happen.
Just the hash of two different nodes in two different genomes could to
hash to the same value. Then no explicit collision happened yet there is
different structure with the same ID.

This is non-detectable to the crossover mechanism which compares/aligns
structure by ID. So crossover will exchange information between
evolutionary unrelated parts, which "makes no sense". But is it harmful?
I assume that depends on the amount of falsely identical structure. So:
How big is the probability of such divergence to propagate? Every
mutation that involves one of the divergent IDs will propagate the
divergence. I'll need to observe the behavior further to inform the
actual impact of this.