Fix bias when sampling start and end nodes
For new feedforward connections the following was the case:
Just chaining the iterators and starting "somewhere" in the iterator
seems like will at least in the long run heavily bias towards sampling
hidden nodes. This is because the amount of hidden nodes can grow while
the number of inputs is fixed. "starting somewhere" is ever more likely
to hit a hidden node, which will then in expectation be followed by
(#hidden nodes / 2) more hidden nodes.

Now the possible nodes are collected and shuffled for uniform sampling.