Fix bias and bug when sampling nodes for recurrent connections
0fb6048329f1bf1089cdab99b1035dbd9bbc3ea0 applied here as well, which is
now fixed. An additional bug was that I chained the output nodes into
the iterator but did not account for the length of output nodes when
taking from the iterator. This resulted in never sampling output nodes
as start nodes for recurrent connections. This is no longer an issue as
all nodes now are collected and iterated.